Move fs mock back to mock file
